Move fleet apply helm test into apply folder
Both apply/ and helm/ are testing the fleet apply sub command
