+++ integrationtests/cli/apply/helm_repository.go
@@ -1,4 +1,4 @@
-package helm
+package apply
 
 import (
 	"crypto/subtle"
